base64_encode: fix input and output length types

+++ proto.h
@@ -829,4 +829,4 @@ void srand48(long);
 long lrand48(void);
 #endif
 
-extern char *base64_encode(const unsigned char *src, int len);
+extern char *base64_encode(const unsigned char *src, size_t len);
